upstream: [package-alt] Sketched interface for package type (#22155)
Port of upstream Sui PR #22155. Introduces the package graph module
(`src/graph/mod.rs`) and moves the `Package` type into its own
`src/package/package_impl.rs`, sketching the package-type interface. Adds a
`Digest` alias in the manifest and `PinnedDependencyInfo` wiring in the local
dependency.
