+++ external-crates/move/crates/move-package-alt/src/cli/build.rs
@@ -20,7 +20,7 @@ impl Build {
     pub async fn execute(&self) -> PackageResult<()> {
         let path = self.path.clone().unwrap_or_else(|| PathBuf::from("."));
 
-        let package = Package::<Vanilla>::load(path, &Vanilla {}).await?;
+        let package = Package::<Vanilla>::load_root(path).await?;
 
         Ok(())
     }
