+++ external-crates/move/crates/move-package-alt/src/lib.rs
@@ -12,5 +12,6 @@ pub mod cli;
 pub mod dependency;
 pub mod errors;
 pub mod flavor;
+pub mod graph;
 pub mod jsonrpc;
 pub mod package;
