+++ external-crates/move/crates/move-package-alt/src/package/manifest.rs
@@ -5,8 +5,8 @@
 
 use std::{
     collections::BTreeMap,
-    fmt,
-    fmt::{Debug, Display, Formatter},
+    fmt::{self, Debug, Display, Formatter},
+    path::Path,
 };
 
 use derive_where::derive_where;
@@ -21,6 +21,9 @@ use crate::{
 
 // TODO: add 2025 edition
 const ALLOWED_EDITIONS: &[&str] = &["2024", "2024.beta", "legacy"];
+
+// TODO: replace this with something more strongly typed
+type Digest = String;
 
 // Note: [Manifest] objects are immutable and should not implement
 // [serde::Serialize]; any tool writing these files should use [toml_edit] to
@@ -158,5 +161,10 @@ impl<F: MoveFlavor> Manifest<F> {
 
     pub fn environments(&self) -> &BTreeMap<EnvironmentName, F::EnvironmentID> {
         &self.environments
+    }
+
+    /// Compute a digest of this file
+    pub fn digest(&self) -> Digest {
+        todo!()
     }
 }
